Fix an error that an user_buffer never gets assigned to a basket.
SetNSlot reads all baskets that contain the first entry before user_buffers gets initialized in TBulkBufferMgr.
In this case, we run into a situation that in GetEntriesSerialized function both basket and user_buffer are created in memory with different addresses.
So we need to make sure they share a same TBuffer in this corner case.